pyproject.toml
[project]
name = "aiagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
